added caching files and loading files from cache

+++ inc/DashboardScraper.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once 'inc/session.php';
+require_once 'inc/files.php';
 
 class DashboardScraper {
     public $urls = array();
@@ -14,19 +15,42 @@ class DashboardScraper {
 
 
         foreach($this->urls as $dashname=>$dashurl) {
+            $basefolder = __DIR__ . '/../cache/dashboards/';
+            $cachefolder = $basefolder . date('YmdH') . '/';
+            $basefolder = 'cache/dashboards/';
+            $cachefolder = $basefolder . date('YmdH') . '/';
+            $filename = $cachefolder . Sanitize(md5($dashurl));
 
-            // get page
-            $curl = curl_init($dashurl);
-            curl_setopt($curl, CURLOPT_RETURNTRANSFER, TRUE);
-            $page = curl_exec($curl);
-            if (curl_errno($curl)) // check for execution errors
-            {
-                AddError('Error while loading url: ' . curl_error($curl));
-                exit;
+            // get page from cache or from url
+            if ( file_exists($filename)) {
+                $_SESSION['debug'][] = 'Getting <a href="' . $dashurl . '">' . $dashurl . '</a> from cache: ' . $filename;
+                // get page from cache
+                $page = file_get_contents($filename);
             }
-            curl_close($curl);
-            $filename = 'cache/' . Sanitize($dashurl);
-            file_put_contents($filename, $page);
+            else
+            {
+                $_SESSION['debug'][] = 'Getting ' . $dashurl . ' from URL';
+                // get page from url, and write to cache
+                $curl = curl_init($dashurl);
+                curl_setopt($curl, CURLOPT_RETURNTRANSFER, TRUE);
+                $page = curl_exec($curl);
+                if (curl_errno($curl)) // check for execution errors
+                {
+                    AddError('Error while loading url: ' . curl_error($curl));
+                    exit;
+                }
+                curl_close($curl);
+
+                // remove old folders from cachefolder
+                rrmdir($basefolder);
+
+                // create timestamped cache folder
+                if (!file_exists($cachefolder)) mkdir($cachefolder, 0777, true);
+
+                // generate random filename and write it (if doesn't exist)
+                if (!file_exists($filename) ) file_put_contents($filename, $page);
+
+            }
 
             // load page into DOM object
             $DOM = new DOMDocument;
